# packages/magic-html-contracts/openapi/tier1.json
{
  "openapi": "3.1.0",
  "info": {"title": "Magic HTML Tier 1 APIs", "version": "1.21.0", "description": "Design Context (Design DNA, direction prose, mobile and desktop canvases, decorative assets), Neo Styler page authoring (a model writes HTML and CSS inside slot, link and judge seat belts), independent styling with immutable Layout Snapshots, signed same-run CDP browser evidence, and finite visual-feedback loops, forms, content, collection, media, illustration, deterministic semantic document templates, replay-safe fixed and generic intake interviews, knowledge retrieval and document lifecycle, tenant-isolated CRM and Messaging, bounded outbound Channel publishing, tenant-safe Measurement and Presence reporting and confirmed writes, isolated remote site-source acquisition, Approval including immutable site-change artifacts, static-site import, safe static-page target inspection and editing, and isolated preview services."},
  "servers": [{"url": "/api"}],
  "paths": {
...
    "/v1/intake-templates/{template}": {
      "put": {
        "operationId": "putGenericIntakeTemplate",
        "security": [{"serviceBearer": []}],
        "parameters": [{"name": "template", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-template-request.json"}}}},
        "responses": {
          "200": {"description": "Existing identical immutable intake definition", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-template-result.json"}}}},
          "201": {"description": "Created immutable intake definition", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-template-result.json"}}}},
          "409": {"description": "The key already represents another definition"}
        }
      }
    },
    "/v1/intake-templates/{template}/sessions": {
      "post": {
        "operationId": "createGenericIntakeSession",
        "security": [{"serviceBearer": []}],
        "parameters": [
          {"name": "template", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 8, "maxLength": 200}}
        ],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-session-request.json"}}}},
        "responses": {
          "200": {"description": "Idempotent replay", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-session-result.json"}}}},
          "201": {"description": "Created intake session and temporary signed URL", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-session-result.json"}}}}
        }
      }
    },
    "/v1/intake-sessions/{interview}": {
      "get": {
        "operationId": "getGenericIntakeSession",
        "security": [{"serviceBearer": []}],
        "parameters": [{"name": "interview", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {"200": {"description": "Intake state and values", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/intake-session-status.json"}}}}}
      }
    },
...
    "/v1/interviews": {
      "post": {
        "tags": ["Interview"],
        "operationId": "startInterview",
        "description": "Start a private scripted interview that accumulates a structured website brief. The key is scoped by site; an exact retry replays the original 201 response and changed input returns 409.",
        "security": [{"serviceBearer": []}],
        "parameters": [{"$ref": "#/components/parameters/InterviewWriteIdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/start-request.json"}}}},
        "responses": {
          "201": {"description": "Active interview session, or the exact idempotent response replayed", "headers": {"Idempotent-Replayed": {"$ref": "#/components/headers/IdempotentReplayed"}}, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/session.json"}}}},
          "401": {"$ref": "#/components/responses/InterviewProblem"},
          "409": {"$ref": "#/components/responses/InterviewProblem"},
          "422": {"$ref": "#/components/responses/InterviewProblem"},
          "429": {"description": "Interview write rate limit exceeded"}
        }
      }
    },
    "/v1/interviews/import": {
      "post": {
        "tags": ["Interview"],
        "operationId": "importInterview",
        "description": "Register an already collected structured website brief as a completed interview session. Idempotency-Key is optional; when supplied, an exact retry returns the existing session and changed input returns 409.",
        "security": [{"serviceBearer": []}],
        "parameters": [{"$ref": "#/components/parameters/InterviewIdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/import-request.json"}}}},
        "responses": {
          "200": {"description": "Existing completed session replayed for the supplied Idempotency-Key", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/session.json"}}}},
          "201": {"description": "Completed interview session created", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/session.json"}}}},
          "401": {"$ref": "#/components/responses/InterviewProblem"},
          "409": {"$ref": "#/components/responses/InterviewProblem"},
          "422": {"$ref": "#/components/responses/InterviewProblem"},
          "429": {"description": "Interview write rate limit exceeded"}
        }
      }
    },
    "/v1/interviews/{interview}": {
      "parameters": [{"$ref": "#/components/parameters/Interview"}],
      "get": {
        "tags": ["Interview"],
        "operationId": "getInterview",
        "description": "Read the messages, current question, and structured data for one private interview session.",
        "security": [{"serviceBearer": []}],
        "responses": {
          "200": {"description": "Interview session", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/session.json"}}}},
          "401": {"$ref": "#/components/responses/InterviewProblem"},
          "404": {"$ref": "#/components/responses/InterviewProblem"},
          "429": {"description": "Interview read rate limit exceeded"}
        }
      },
      "delete": {
        "tags": ["Interview"],
        "operationId": "deleteInterview",
        "description": "Delete one private interview authoring session.",
        "security": [{"serviceBearer": []}],
        "responses": {
          "204": {"description": "Interview session deleted"},
          "401": {"$ref": "#/components/responses/InterviewProblem"},
          "404": {"$ref": "#/components/responses/InterviewProblem"},
          "429": {"description": "Interview write rate limit exceeded"}
        }
      }
    },
    "/v1/interviews/{interview}/messages": {
      "parameters": [{"$ref": "#/components/parameters/Interview"}],
      "post": {
        "tags": ["Interview"],
        "operationId": "answerInterview",
        "description": "Answer the active session's expected current scripted step and advance its structured brief. The key is scoped by site and session; exact retries replay the original response, while changed input, concurrent work, stale steps, and completed sessions return 409 without advancing twice.",
        "security": [{"serviceBearer": []}],
        "parameters": [{"$ref": "#/components/parameters/InterviewWriteIdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/answer-request.json"}}}},
        "responses": {
          "200": {"description": "Advanced interview session, or the exact idempotent response replayed", "headers": {"Idempotent-Replayed": {"$ref": "#/components/headers/IdempotentReplayed"}}, "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/session.json"}}}},
          "401": {"$ref": "#/components/responses/InterviewProblem"},
          "404": {"$ref": "#/components/responses/InterviewProblem"},
          "409": {"$ref": "#/components/responses/InterviewProblem"},
          "422": {"$ref": "#/components/responses/InterviewProblem"},
          "429": {"description": "Interview write rate limit exceeded"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceBearer": {"type": "http", "scheme": "bearer", "description": "A trusted service credential. The caller authenticates and authorizes end users for tenant/site scope and asserts actor identity."},
      "slackHmacV0": {"type": "apiKey", "in": "header", "name": "X-Slack-Signature", "description": "Slack v0 HMAC-SHA256 of v0:timestamp:exact_raw_body using the channel's runtime-resolved signing secret. X-Slack-Request-Timestamp is also required and limited to a 300-second skew."}
    },
    "parameters": {
...
      "Interview": {"name": "interview", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "InterviewIdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": false, "description": "Optional import idempotency key. Exact retries return the existing completed session; reuse for changed input returns 409.", "schema": {"type": "string", "minLength": 8, "maxLength": 200}},
      "InterviewWriteIdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "description": "Opaque start or answer key scoped to the relevant site and, for answers, interview session. Exact retries replay the original status and body; changed input or work already in progress returns 409.", "schema": {"type": "string", "minLength": 8, "maxLength": 200}},
...
    },
    "headers": {
      "IdempotentReplayed": {"description": "Whether this response is an exact replay of a completed operation.", "schema": {"type": "string", "enum": ["true", "false"]}}
    },
    "responses": {
...
      "InterviewProblem": {"description": "Interview API problem", "content": {"application/json": {"schema": {"$ref": "../schemas/v1/interview/problem.json"}}}},
...
    }
  }
}